Extract UserObservationCollection to models package
